+++ projectroles/templatetags/projectroles_common_tags.py
@@ -95,7 +95,7 @@ def get_django_setting(name, js=False):
 # DEPRECATION PROTECTION: To be removed in v0.7.0, use get_django_setting()
 @register.simple_tag
 def get_setting(name, js=False):
-    """bb
+    """
     Return value of Django setting by name or None if it is not found.
     Return a Javascript-safe value if js=True.
     """